Add files via upload
This is a non GUI AI interactive game coded in Java. In my next project I'm going to create the same Tic Tac Toe game but I'm going to add GUI with JavaFX.
This is my first lone project I have created. I'm open for suggestions or changes to my code to make it better!